requirements.txt


async-timeout==4.0.2
attrs==22.1.0
autopep8==1.6.0
certifi==2022.9.24
charset-normalizer==2.1.1
discord==2.0.0
discord.py @ git+https://github.com/Rapptz/discord.py@7b595fb052c82de1fca8b89d6aacc908f1978a31
frozenlist==1.3.1
idna==3.4
multidict==6.0.2
pycodestyle==2.8.0
python-dotenv==0.21.0
toml==0.10.2
yarl==1.8.1
